add html post submit button; change to post api mode; let the service restart itself when down; let ie6 use get request

+++ yppm/resources/http_python_message_board/main.py
@@ -17,19 +17,19 @@ def start_work_process():
     if 'work' in sys.modules:
         del sys.modules['work']
 
-    import work  # Now import will reflect the latest changes
-
     global work_process
 
     while work_process == None or work_process.is_alive() == False:
         try:
+            import work  # Now import will reflect the latest changes
             http_port = 8899
             work_process = Process(target=work.work_function, args=(http_port, ))
             work_process.start()
-            time.sleep(10)
         except Exception as e:
-            print(e)
-            work_process.terminate()
+            print("error:", e)
+            if work_process != None:
+                work_process.terminate()
+        time.sleep(10)
 
     #work_process.join()
 
@@ -42,10 +42,10 @@ def restart_work_process():
         print("process killed.\n")
         while work_process.is_alive() == True:
             time.sleep(1)
-        start_work_process()
+    start_work_process()
 
 
-start_work_process()
+restart_work_process()
 
 # detect_file_changes, if changed, restart the work process
 main_program_file_path = "./work.py"
@@ -56,4 +56,8 @@ while True:
         print("File has changed!")
         last_modified = current_modified
         restart_work_process()
+    if work_process != None:
+        if work_process.is_alive() == False:
+            restart_work_process()
+            time.sleep(10)
     time.sleep(1) # 1 second
